various changes
- note/text rendering jitter fix
- added decimal support for text box settings
- selection hitbox optimization
- switched export button to hidden keybind (ss+ management fell apart practically overnight)
- added debug message for possible save bug, wont remain in release
- adjustment to autosave management
- added failsafe error if editor fails to start without otherwise erroring
- fixed switching maps breaking functions

+++ Blox Saber Editor/Gui/Gui.cs
@@ -220,7 +220,7 @@ namespace Sound_Space_Editor.GUI
         {
             var renderer = MainWindow.Instance.Fonts[font];
 
-            renderer.Render(text, (int)posx, (int)posy, size);
+            renderer.Render(text, posx, posy, size);
         }
 
         public int TextWidth(string text, int size, string font = "main")
